requirements.txt
fastapi
uvicorn
gunicorn
requests
openpyxl
python-dotenv
psycopg2-binary
anthropic
pandas
